[Property] Expose the UPnP createClass property
Fix https://github.com/01org/media-service-upnp/issues/45

Expose the UPnP createClasses property on containers.
This property indicates what object types can be created in a container.
Upload controllers need this information so that they can correctly display
the valid options to users.

New property "CreateClasses" is added to org.gnome.MediaContainer2.
It exposes the UPnP createClass property of that folder.
"CreateClasses" property is an array of tuples: a(sb).
Each tuble (sb) represents the class and whether derived values
are allowed for this class.

API.txt file has been updated to reflect the addition of this new property.

Signed-off-by: Christophe Guiraud <christophe.guiraud@intel.com>