[Upload] Implemented GetUploadIDs method
The GetUploadIDs method allows clients to get a list of the IDs of
all the outstanding uploads that have been initiated from an instance of
media-service-upnp.  These IDs can be used to retrieve information about
the status of the upload.  In addition, it will be possible in later
versions of media-service-upnp to use these ids to cancel uploads.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>